Add FitImage to example app
