+++ FitImageExample/index.ios.js
@@ -1,53 +1,16 @@
 /**
- * Sample React Native App
- * https://github.com/facebook/react-native
- * @flow
+ * Example for FitImage
+ * https://github.com/originerd/react-native-fit-image
  */
 
 import React, {
   AppRegistry,
-  Component,
-  StyleSheet,
-  Text,
-  View
 } from 'react-native';
 
-class FitImageExample extends Component {
-  render() {
-    return (
-      <View style={styles.container}>
-        <Text style={styles.welcome}>
-          Welcome to React Native!
-        </Text>
-        <Text style={styles.instructions}>
-          To get started, edit index.ios.js
-        </Text>
-        <Text style={styles.instructions}>
-          Press Cmd+R to reload,{'\n'}
-          Cmd+D or shake for dev menu
-        </Text>
-      </View>
-    );
-  }
-}
+import App from './src/app/App';
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#F5FCFF',
-  },
-  welcome: {
-    fontSize: 20,
-    textAlign: 'center',
-    margin: 10,
-  },
-  instructions: {
-    textAlign: 'center',
-    color: '#333333',
-    marginBottom: 5,
-  },
-});
+function FitImageExample() {
+  return <App />;
+}
 
 AppRegistry.registerComponent('FitImageExample', () => FitImageExample);
