+++ FitImageExample/package.json
@@ -7,6 +7,7 @@
   },
   "dependencies": {
     "react": "^0.14.8",
-    "react-native": "^0.24.1"
+    "react-native": "^0.24.1",
+    "react-native-fit-image": "^1.0.0"
   }
 }
